+++ packages/components/src/grid/hook.js
@@ -18,7 +18,7 @@ import CONFIG from '../utils/config-values';
 import { useCx } from '../utils/hooks/use-cx';
 
 /**
- * @param {import('../ui/context').PolymorphicComponentProps<import('./types').Props, 'div'>} props
+ * @param {import('../ui/context').WordPressComponentProps<import('./types').Props, 'div'>} props
  */
 export default function useGrid( props ) {
 	const {