+++ packages/core-data/src/test/actions.js
@@ -1,7 +1,6 @@
 /**
  * WordPress dependencies
  */
-import { controls } from '@wordpress/data';
 import apiFetch from '@wordpress/api-fetch';
 
 jest.mock( '@wordpress/api-fetch' );
@@ -19,19 +18,6 @@ import {
 	__experimentalBatch,
 } from '../actions';
 
-jest.mock( '../locks/actions', () => ( {
-	__unstableAcquireStoreLock: jest.fn( () => [
-		{
-			type: 'MOCKED_ACQUIRE_LOCK',
-		},
-	] ),
-	__unstableReleaseStoreLock: jest.fn( () => [
-		{
-			type: 'MOCKED_RELEASE_LOCK',
-		},
-	] ),
-} ) );
-
 jest.mock( '../batch', () => {
 	const { createBatch } = jest.requireActual( '../batch' );
 	return {
@@ -42,68 +28,81 @@ jest.mock( '../batch', () => {
 } );
 
 describe( 'editEntityRecord', () => {
-	it( 'throws when the edited entity does not have a loaded config.', () => {
+	it( 'throws when the edited entity does not have a loaded config.', async () => {
 		const entity = { kind: 'someKind', name: 'someName', id: 'someId' };
+		const select = {
+			getEntity: jest.fn(),
+		};
 		const fulfillment = editEntityRecord(
 			entity.kind,
 			entity.name,
 			entity.id,
 			{}
-		);
-		expect( fulfillment.next().value ).toEqual(
-			controls.select( 'core', 'getEntity', entity.kind, entity.name )
-		);
-
-		// Don't pass back an entity config.
-		expect( fulfillment.next.bind( fulfillment ) ).toThrow(
+		)( { select } );
+		expect( select.getEntity ).toHaveBeenCalledTimes( 1 );
+		await expect( fulfillment ).rejects.toThrow(
 			`The entity being edited (${ entity.kind }, ${ entity.name }) does not have a loaded config.`
 		);
 	} );
 } );
 
 describe( 'deleteEntityRecord', () => {
+	beforeEach( async () => {
+		apiFetch.mockReset();
+		jest.useFakeTimers();
+	} );
+
 	it( 'triggers a DELETE request for an existing record', async () => {
-		const post = 10;
+		const deletedRecord = { title: 'new post', id: 10 };
 		const entities = [
 			{ name: 'post', kind: 'postType', baseURL: '/wp/v2/posts' },
 		];
-		const fulfillment = deleteEntityRecord( 'postType', 'post', post );
 
-		// Trigger generator
-		fulfillment.next();
+		const dispatch = Object.assign( jest.fn(), {
+			receiveEntityRecords: jest.fn(),
+			__unstableAcquireStoreLock: jest.fn(),
+			__unstableReleaseStoreLock: jest.fn(),
+		} );
+		// Provide entities
+		dispatch.mockReturnValueOnce( entities );
 
-		// Acquire lock
-		expect( fulfillment.next( entities ).value.type ).toBe(
-			'MOCKED_ACQUIRE_LOCK'
-		);
+		// Provide response
+		apiFetch.mockImplementation( () => deletedRecord );
 
-		// Start
-		expect( fulfillment.next().value.type ).toEqual(
-			'DELETE_ENTITY_RECORD_START'
-		);
+		const result = await deleteEntityRecord(
+			'postType',
+			'post',
+			deletedRecord.id
+		)( { dispatch } );
 
-		// delete api call
-		const { value: apiFetchAction } = fulfillment.next();
-		expect( apiFetchAction.request ).toEqual( {
+		expect( apiFetch ).toHaveBeenCalledTimes( 1 );
+		expect( apiFetch ).toHaveBeenCalledWith( {
 			path: '/wp/v2/posts/10',
 			method: 'DELETE',
 		} );
 
-		expect( fulfillment.next().value.type ).toBe( 'REMOVE_ITEMS' );
-
-		expect( fulfillment.next().value.type ).toBe(
-			'DELETE_ENTITY_RECORD_FINISH'
-		);
-
-		// Release lock
-		expect( fulfillment.next().value.type ).toEqual(
-			'MOCKED_RELEASE_LOCK'
-		);
-
-		expect( fulfillment.next() ).toMatchObject( {
-			done: true,
-			value: undefined,
+		expect( dispatch ).toHaveBeenCalledTimes( 4 );
+		expect( dispatch ).toHaveBeenCalledWith( {
+			type: 'DELETE_ENTITY_RECORD_START',
+			kind: 'postType',
+			name: 'post',
+			recordId: 10,
 		} );
+		expect( dispatch ).toHaveBeenCalledWith( {
+			type: 'DELETE_ENTITY_RECORD_FINISH',
+			kind: 'postType',
+			name: 'post',
+			recordId: 10,
+			error: undefined,
+		} );
+		expect( dispatch.__unstableAcquireStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
+		expect( dispatch.__unstableReleaseStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
+
+		expect( result ).toBe( deletedRecord );
 	} );
 } );
 
@@ -124,6 +123,8 @@ describe( 'saveEntityRecord', () => {
 
 		const dispatch = Object.assign( jest.fn(), {
 			receiveEntityRecords: jest.fn(),
+			__unstableAcquireStoreLock: jest.fn(),
+			__unstableReleaseStoreLock: jest.fn(),
 		} );
 		// Provide entities
 		dispatch.mockReturnValueOnce( entities );
@@ -147,7 +148,7 @@ describe( 'saveEntityRecord', () => {
 			data: post,
 		} );
 
-		expect( dispatch ).toHaveBeenCalledTimes( 5 );
+		expect( dispatch ).toHaveBeenCalledTimes( 3 );
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_START',
 			kind: 'postType',
@@ -155,11 +156,9 @@ describe( 'saveEntityRecord', () => {
 			recordId: undefined,
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_ACQUIRE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableAcquireStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_FINISH',
 			kind: 'postType',
@@ -168,11 +167,9 @@ describe( 'saveEntityRecord', () => {
 			error: undefined,
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_RELEASE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableReleaseStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledTimes( 1 );
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledWith(
@@ -198,6 +195,8 @@ describe( 'saveEntityRecord', () => {
 
 		const dispatch = Object.assign( jest.fn(), {
 			receiveEntityRecords: jest.fn(),
+			__unstableAcquireStoreLock: jest.fn(),
+			__unstableReleaseStoreLock: jest.fn(),
 		} );
 		// Provide entities
 		dispatch.mockReturnValueOnce( entities );
@@ -221,7 +220,7 @@ describe( 'saveEntityRecord', () => {
 			data: post,
 		} );
 
-		expect( dispatch ).toHaveBeenCalledTimes( 5 );
+		expect( dispatch ).toHaveBeenCalledTimes( 3 );
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_START',
 			kind: 'postType',
@@ -229,11 +228,9 @@ describe( 'saveEntityRecord', () => {
 			recordId: 10,
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_ACQUIRE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableAcquireStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_FINISH',
 			kind: 'postType',
@@ -242,11 +239,9 @@ describe( 'saveEntityRecord', () => {
 			error: undefined,
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_RELEASE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableReleaseStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledTimes( 1 );
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledWith(
@@ -277,6 +272,8 @@ describe( 'saveEntityRecord', () => {
 
 		const dispatch = Object.assign( jest.fn(), {
 			receiveEntityRecords: jest.fn(),
+			__unstableAcquireStoreLock: jest.fn(),
+			__unstableReleaseStoreLock: jest.fn(),
 		} );
 		// Provide entities
 		dispatch.mockReturnValueOnce( entities );
@@ -297,7 +294,7 @@ describe( 'saveEntityRecord', () => {
 			data: postType,
 		} );
 
-		expect( dispatch ).toHaveBeenCalledTimes( 5 );
+		expect( dispatch ).toHaveBeenCalledTimes( 3 );
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_START',
 			kind: 'root',
@@ -305,11 +302,9 @@ describe( 'saveEntityRecord', () => {
 			recordId: 'page',
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_ACQUIRE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableAcquireStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 		expect( dispatch ).toHaveBeenCalledWith( {
 			type: 'SAVE_ENTITY_RECORD_FINISH',
 			kind: 'root',
@@ -318,11 +313,9 @@ describe( 'saveEntityRecord', () => {
 			error: undefined,
 			isAutosave: false,
 		} );
-		expect( dispatch ).toHaveBeenCalledWith( [
-			{
-				type: 'MOCKED_RELEASE_LOCK',
-			},
-		] );
+		expect( dispatch.__unstableReleaseStoreLock ).toHaveBeenCalledTimes(
+			1
+		);
 
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledTimes( 1 );
 		expect( dispatch.receiveEntityRecords ).toHaveBeenCalledWith(